+++ lib/presentation/page/setting/setting_page.dart
@@ -5,6 +5,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../application/url_launcher/url_launcher_service.dart';
 import '../../../domain/entity/app_info.dart';
 import '../../component/delete_user.dart';
+import '../../component/list_tile.dart';
 import '../../router.dart';
 
 /// 設定画面
@@ -30,17 +31,17 @@ class _Body extends StatelessWidget {
     return SingleChildScrollView(
       child: Column(
         children: const [
-          _SectionHeader(title: 'アカウント'),
+          SectionHeader(title: 'アカウント'),
           _DeleteUserItem(),
 
-          _SectionHeader(title: 'サポート'),
+          SectionHeader(title: 'サポート'),
           _TermsOfServiceItem(),
           _PrivacyPolicyItem(),
           _AboutAppItem(),
 
           // デバッグモードのときだけ表示する
           if (kDebugMode) ...[
-            _SectionHeader(title: 'デバッグ'),
+            SectionHeader(title: 'デバッグ'),
             _ComponentGalleryItem(),
           ],
         ],
@@ -55,7 +56,7 @@ class _DeleteUserItem extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return _SectionItem(
+    return SectionItem(
       onTap: () => showDialog<void>(
         context: context,
         builder: (_) => const DeleteUserConfirmDialog(),
@@ -76,7 +77,7 @@ class _TermsOfServiceItem extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return _SectionItem(
+    return SectionItem(
       title: const Text('サービス利用規約について'),
       onTap: () async {
         await ref.read(urlLauncherServiceProvider).launch(
@@ -93,7 +94,7 @@ class _PrivacyPolicyItem extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return _SectionItem(
+    return SectionItem(
       title: const Text('プライバシーポリシー'),
       onTap: () async {
         await ref.read(urlLauncherServiceProvider).launch(
@@ -111,7 +112,7 @@ class _AboutAppItem extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final info = ref.watch(appInfoProvider);
-    return _SectionItem(
+    return SectionItem(
       title: const Text('このアプリについて'),
       subtitle: Text(info.version),
       onTap: () {
@@ -133,74 +134,15 @@ class _AboutAppItem extends ConsumerWidget {
   }
 }
 
-/// 区切り線の高さ
-const _dividerHeight = 1.0;
-
-/// セクションヘッダー
-class _SectionHeader extends StatelessWidget {
-  const _SectionHeader({
-    required this.title,
-  });
-
-  final String title;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        ListTile(
-          title: Text(
-            title,
-            style: const TextStyle(fontWeight: FontWeight.bold),
-          ),
-        ),
-        const Divider(
-          height: _dividerHeight,
-        ),
-      ],
-    );
-  }
-}
-
 /// Component Gallery
 class _ComponentGalleryItem extends ConsumerWidget {
   const _ComponentGalleryItem();
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return _SectionItem(
+    return SectionItem(
       title: const Text(ComponentGalleryRoute.title),
       onTap: () => const ComponentGalleryRoute().go(context),
-    );
-  }
-}
-
-/// セクション項目
-class _SectionItem extends StatelessWidget {
-  const _SectionItem({
-    required this.title,
-    this.subtitle,
-    this.onTap,
-  });
-
-  final Widget? title;
-  final Widget? subtitle;
-  final GestureTapCallback? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        ListTile(
-          title: title,
-          subtitle: subtitle,
-          onTap: onTap,
-          tileColor: Theme.of(context).colorScheme.surface,
-        ),
-        const Divider(
-          height: _dividerHeight,
-        ),
-      ],
     );
   }
 }
